services/chat-service/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyJWT==2.8.0
redis==5.0.1
pymongo==4.6.0
boto3==1.34.0
python-dotenv==1.0.0
eventlet==0.33.3 